+++ src/defs.h
@@ -119,6 +119,7 @@ typedef struct {
 	Bool enabled;
 } Scratchpad;
 
+extern void centre_window();
 extern void close_focused(void);
 extern void dec_gaps(void);
 extern void focus_next(void);